backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai==1.97.1
python-dotenv
geopandas
shapely
firebase-admin
requests
matplotlib
boto3